KOA-4605 Fix icon alignment
